+++ extension/components/queryComponents/RecentTermsToolTip.js
